_pipelines/requirements.txt
apache-airflow~=2.8.0
requests~=2.28.2
